merge

+++ src/main/java/repository/MySQLClienteDAO.java
@@ -1,14 +1,27 @@
 package repository;
 import entity.Cliente;
 import dao.ClienteDAO;
+import entity.Producto;
 
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Implementación de acceso a datos para la entidad {@link Cliente} utilizando MySQL.
+ * Provee operaciones CRUD básicas y consultas de agregación.
+ *
+ *
+ * @version 1.0
+ */
 
 public class MySQLClienteDAO implements ClienteDAO {
     private final Connection cn;
+    /**
+     * Construye una nueva instancia del DAO y crea la tabla si no existe.
+     *
+     * @param cn Conexión activa a la base de datos MySQL.
+     */
 
     public MySQLClienteDAO(Connection cn) {
         this.cn = cn;
@@ -116,7 +129,7 @@ public class MySQLClienteDAO implements ClienteDAO {
     public void delete(int id) {
         final String sql = "DELETE FROM clientes WHERE idCliente = ?";
         try (PreparedStatement ps = cn.prepareStatement(sql)) {
-            ps.setLong(1, id);
+            ps.setInt(1, id);
             ps.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException("Error en delete", e);
